refactored to consildate policy reset and job hash into a single file
