+++ src/marine-licences/api/helpers/marine-plan-policies/policy-job.js
@@ -44,3 +44,20 @@ export const computePolicyJobId = (licenceId, siteDetails = []) => {
     .update(`${licenceId}:${sortedGeometries.join('|')}`)
     .digest('hex')
 }
+
+// marinePlanPolicyResponses are deliberately never reset — only policy job state is discarded on geometry change.
+export const buildPolicyResetFields = (id, existing, newSiteDetails) => {
+  if (!existing?.marinePlanPolicyJobId) {
+    return {}
+  }
+  const newPolicyJobId = computePolicyJobId(id, newSiteDetails)
+  if (existing.marinePlanPolicyJobId === newPolicyJobId) {
+    return {}
+  }
+  return {
+    marinePlanPolicyJob: null,
+    marinePlanPolicyJobId: null,
+    marinePlanPolicyJobQueuedAt: null,
+    marinePlanPolicies: []
+  }
+}
